Refactor: Separate snapshot_tracker class into standalone files
